+++ client/src/app/shared/directives/resizable.directive.ts
@@ -1,7 +1,5 @@
 import { Directive, ElementRef, Input, OnInit } from '@angular/core';
 
-import * as $ from 'jquery';
-import 'jqueryui';
 import { isNil } from 'lodash-es';
 
 
@@ -25,12 +23,12 @@ export class ResizableDirective  implements OnInit {
       return;
     }
 
-    $(`#${this.el.nativeElement.id}`).resizable({
-      handles: this.handles,
-      maxWidth: 500,
-      minWidth: 256,
-      maxHeight: 500,
-      minHeight: this.minHeight
-    });
+    const element: HTMLElement = this.el.nativeElement;
+    element.style.resize = 'both';
+    element.style.overflow = 'auto';
+    element.style.maxWidth = '500px';
+    element.style.minWidth = '256px';
+    element.style.maxHeight = '500px';
+    element.style.minHeight = `${this.minHeight}px`;
   }
 }